fix Model::softmax()で指数関数の返戻値をfloatに修正

+++ src/Model.cpp
@@ -37,7 +37,7 @@ void Model::softmax(const vector<vector<float>> &outputs,
         sum_exp = 0.f;
         max_output = (*std::max_element(outputs[i].begin(), outputs[i].end()));
         for (int j = 0; j < outputs[i].size(); ++j) {
-            u = exp(outputs[i][j] - max_output);
+            u = expf(outputs[i][j] - max_output);
             y[i][j] = u;
             sum_exp += u;
         }
